fix: sanitize rST feed HTML and stop mis-parsing rST fallback as markdown
Full-content feeds embedded docutils renderedHtml verbatim while the
on-page RstRenderer sanitizes it, and when docutils was unavailable the
rST *source* went through the Markdown parser, leaking literal
.. directive:: text into <content:encoded>.

Hoist the sanitize-html config out of RstRenderer.tsx into
src/lib/rst-sanitize.ts so both consumers share one allowlist (the
Shiki-attr and radio-tab decisions are load-bearing and must not
drift), sanitize the feed's rST HTML with it, and fall back to the
plain-text excerpt when renderedHtml is unset. New tests cover both
paths behind the same environment gates as the existing rST parity
tests.
